Move Streams/Parallel.hs to Internal/Data/Stream
