+++ src/Streamly.hs
@@ -227,7 +227,7 @@ import Streamly.Internal.Data.SVar (MonadAsync, Rate(..))
 import Streamly.Internal.Data.Stream.Ahead
 import Streamly.Internal.Data.Stream.Async
 import Streamly.Internal.Data.Stream.Combinators
-import Streamly.Streams.Parallel
+import Streamly.Internal.Data.Stream.Parallel
 import Streamly.Internal.Data.Stream.Serial
 import Streamly.Internal.Data.Stream.StreamK hiding (serial)
 import Streamly.Internal.Data.Stream.Zip
